requirements.txt
transformers
sentence-transformers
faiss-cpu
langchain
openai
pandas
numpy